Refactors RSS data fetching logic
Moves RSS data fetching from Home component to RssFeedMetadata component
Adds logger and URL as props to RssFeedMetadata for better separation of concerns
